+++ rust/operator-binary/src/hdfs_controller.rs
@@ -48,6 +48,7 @@ use stackable_operator::{
         rollout::check_statefulset_rollout_complete,
     },
     time::Duration,
+    utils::cluster_info::KubernetesClusterInfo,
 };
 use strum::{EnumDiscriminants, IntoEnumIterator, IntoStaticStr};
 
@@ -412,6 +413,7 @@ pub async fn reconcile_hdfs(
 
             let rg_configmap = rolegroup_config_map(
                 hdfs,
+                &client.kubernetes_cluster_info,
                 metadata,
                 &rolegroup_ref,
                 rolegroup_config,
@@ -424,6 +426,7 @@ pub async fn reconcile_hdfs(
 
             let rg_statefulset = rolegroup_statefulset(
                 hdfs,
+                &client.kubernetes_cluster_info,
                 metadata,
                 &role,
                 &rolegroup_ref,
@@ -485,6 +488,7 @@ pub async fn reconcile_hdfs(
     // so that failure won't inhibit the rest of the cluster from booting up.
     let discovery_cm = build_discovery_configmap(
         hdfs,
+        &client.kubernetes_cluster_info,
         HDFS_CONTROLLER,
         &hdfs
             .namenode_listener_refs(client)
@@ -601,6 +605,7 @@ fn rolegroup_service(
 #[allow(clippy::too_many_arguments)]
 fn rolegroup_config_map(
     hdfs: &HdfsCluster,
+    cluster_info: &KubernetesClusterInfo,
     metadata: &ObjectMetaBuilder,
     rolegroup_ref: &RoleGroupRef<HdfsCluster>,
     rolegroup_config: &HashMap<PropertyNameKind, BTreeMap<String, String>>,
@@ -652,10 +657,10 @@ fn rolegroup_config_map(
                     .dfs_replication(hdfs.spec.cluster_config.dfs_replication)
                     .dfs_name_services()
                     .dfs_ha_namenodes(namenode_podrefs)
-                    .dfs_namenode_shared_edits_dir(journalnode_podrefs)
+                    .dfs_namenode_shared_edits_dir(cluster_info, journalnode_podrefs)
                     .dfs_namenode_name_dir_ha(namenode_podrefs)
-                    .dfs_namenode_rpc_address_ha(namenode_podrefs)
-                    .dfs_namenode_http_address_ha(hdfs, namenode_podrefs)
+                    .dfs_namenode_rpc_address_ha(cluster_info, namenode_podrefs)
+                    .dfs_namenode_http_address_ha(hdfs, cluster_info, namenode_podrefs)
                     .dfs_client_failover_proxy_provider()
                     .security_config(hdfs)
                     .add("dfs.ha.fencing.methods", "shell(/bin/true)")
@@ -684,7 +689,7 @@ fn rolegroup_config_map(
                 core_site
                     .fs_default_fs()
                     .ha_zookeeper_quorum()
-                    .security_config(hdfs)
+                    .security_config(hdfs, cluster_info)
                     .context(BuildSecurityConfigSnafu)?;
                 if let Some(hdfs_opa_config) = hdfs_opa_config {
                     hdfs_opa_config.add_core_site_config(&mut core_site);
@@ -805,6 +810,7 @@ fn rolegroup_config_map(
 #[allow(clippy::too_many_arguments)]
 fn rolegroup_statefulset(
     hdfs: &HdfsCluster,
+    cluster_info: &KubernetesClusterInfo,
     metadata: &ObjectMetaBuilder,
     role: &HdfsRole,
     rolegroup_ref: &RoleGroupRef<HdfsCluster>,
@@ -849,6 +855,7 @@ fn rolegroup_statefulset(
     ContainerConfig::add_containers_and_volumes(
         &mut pb,
         hdfs,
+        cluster_info,
         role,
         resolved_product_image,
         merged_config,
@@ -911,6 +918,8 @@ pub fn error_policy(
 
 #[cfg(test)]
 mod test {
+    use stackable_operator::commons::networking::DomainName;
+
     use super::*;
 
     #[test]
@@ -983,6 +992,9 @@ properties: []
         ContainerConfig::add_containers_and_volumes(
             &mut pb,
             &hdfs,
+            &KubernetesClusterInfo {
+                cluster_domain: DomainName::try_from("cluster.local").unwrap(),
+            },
             &role,
             &resolved_product_image,
             &merged_config,